wireguard code restructure wireguard addpeer bug fix
