+++ ToDo
@@ -1,5 +1,4 @@
 4 - implement wireguard
-TEST
 last handshake
 Tx and Rx
