+++ backend/bridge/handlers/wireguard/wireguard.go
@@ -2,17 +2,14 @@ package wireguard
 
 import (
 	"encoding/base64"
-	"encoding/binary"
-	"encoding/json"
 	"fmt"
-	"net"
 	"os"
 	"os/exec"
 	"path/filepath"
 	"strconv"
 	"strings"
 
-	qrcode "github.com/skip2/go-qrcode"
+	"github.com/skip2/go-qrcode"
 	"golang.zx2c4.com/wireguard/wgctrl/wgtypes"
 	"gopkg.in/ini.v1"
 
@@ -21,350 +18,14 @@ import (
 	"github.com/mordilloSan/LinuxIO/internal/utils"
 )
 
-// --- Constants ---
-const (
-	wgConfigDir      = "/etc/wireguard"
-	configExt        = ".conf"
-	defaultKeepalive = 25
-	maxHostIP        = 254
-	minHostIP        = 2
-)
-
-// --- Types ---
-type InterfaceConfig struct {
-	PrivateKey string           `json:"private_key"`
-	Address    []string         `json:"address"`
-	ListenPort int              `json:"listen_port"`
-	DNS        []string         `json:"dns"`
-	MTU        int              `json:"mtu"`
-	Peers      []ipc.PeerConfig `json:"peers"`
-}
-
-type WireGuardInterfaceUI struct {
-	Name        string `json:"name"`
-	IsConnected string `json:"isConnected"`
-	Address     string `json:"address"`
-	Port        int    `json:"port"`
-	PeerCount   int    `json:"peerCount"`
-}
-
-// --- Path Helpers ---
-func configPath(name string) string {
-	return filepath.Join(wgConfigDir, name+configExt)
-}
-
-func peerDirPath(iface string) string {
-	return filepath.Join(wgConfigDir, iface)
-}
-
-func peerConfigPath(iface, peerName string) string {
-	return filepath.Join(peerDirPath(iface), peerName+configExt)
-}
-
-// --- Config Parsing ---
-func ParseWireGuardConfig(path string) (InterfaceConfig, error) {
-	var cfg InterfaceConfig
-
-	iniFile, err := ini.LoadSources(ini.LoadOptions{
-		AllowNonUniqueSections: true,
-	}, path)
-	if err != nil {
-		return cfg, fmt.Errorf("load config: %w", err)
-	}
-
-	// Parse Interface section
-	ifSec := iniFile.Section("Interface")
-	cfg.PrivateKey = ifSec.Key("PrivateKey").String()
-	cfg.Address = parseCSV(ifSec.Key("Address").String())
-	cfg.ListenPort, _ = ifSec.Key("ListenPort").Int()
-	cfg.DNS = parseCSV(ifSec.Key("DNS").String())
-	cfg.MTU, _ = ifSec.Key("MTU").Int()
-
-	// Parse Peer sections
-	peerIdx := 1
-	for _, sec := range iniFile.Sections() {
-		if !isPeerSection(sec.Name()) {
-			continue
-		}
-
-		pc := ipc.PeerConfig{
-			PublicKey:    sec.Key("PublicKey").String(),
-			PresharedKey: sec.Key("PresharedKey").String(),
-			Endpoint:     sec.Key("Endpoint").String(),
-			Name:         sec.Key("Name").String(),
-			AllowedIPs:   parseCSV(sec.Key("AllowedIPs").String()),
-		}
-
-		if pc.Name == "" {
-			pc.Name = fmt.Sprintf("peer%d", peerIdx)
-		}
-
-		pc.PersistentKeepalive, _ = sec.Key("PersistentKeepalive").Int()
-		cfg.Peers = append(cfg.Peers, pc)
-		peerIdx++
-	}
-
-	return cfg, nil
-}
-
-func WriteWireGuardConfig(path string, cfg InterfaceConfig) error {
-	return writeConfig(path, cfg, "", false)
-}
-
-func WriteWireGuardConfigWithPostUpDown(path string, cfg InterfaceConfig, egressNic string) error {
-	return writeConfig(path, cfg, egressNic, true)
-}
-
-// Unified config writer
-func writeConfig(path string, cfg InterfaceConfig, egressNic string, includePostUpDown bool) error {
-	iniFile := ini.Empty(ini.LoadOptions{AllowNonUniqueSections: true})
-
-	// Create Interface section
-	ifSec, err := iniFile.NewSection("Interface")
-	if err != nil {
-		return fmt.Errorf("create interface section: %w", err)
-	}
-
-	// Set interface keys
-	setKeyIfNotEmpty(ifSec, "Address", strings.Join(cfg.Address, ","))
-	setKeyIfPositive(ifSec, "ListenPort", cfg.ListenPort)
-	setKey(ifSec, "PrivateKey", cfg.PrivateKey)
-	setKeyIfNotEmpty(ifSec, "DNS", strings.Join(cfg.DNS, ","))
-	setKeyIfPositive(ifSec, "MTU", cfg.MTU)
-
-	// Add PostUp/PostDown for NAT if requested
-	if includePostUpDown && egressNic != "" {
-		// Get the subnet from the first address
-		subnet := cfg.Address[0]
-
-		postUp := "sysctl -w net.ipv4.ip_forward=1; "
-		postUp += fmt.Sprintf("iptables -I FORWARD 1 -i %%i -o %s -j ACCEPT; ", egressNic)
-		postUp += fmt.Sprintf("iptables -I FORWARD 1 -o %%i -i %s -m state --state RELATED,ESTABLISHED -j ACCEPT; ", egressNic)
-		postUp += fmt.Sprintf("iptables -t nat -A POSTROUTING -o %s -s %s -j MASQUERADE", egressNic, subnet)
-
-		postDown := fmt.Sprintf("iptables -D FORWARD -i %%i -o %s -j ACCEPT; ", egressNic)
-		postDown += fmt.Sprintf("iptables -D FORWARD -o %%i -i %s -m state --state RELATED,ESTABLISHED -j ACCEPT; ", egressNic)
-		postDown += fmt.Sprintf("iptables -t nat -D POSTROUTING -o %s -s %s -j MASQUERADE", egressNic, subnet)
-
-		setKey(ifSec, "PostUp", postUp)
-		setKey(ifSec, "PostDown", postDown)
-	}
-
-	// Create Peer sections
-	for _, peer := range cfg.Peers {
-		if err := addPeerSection(iniFile, peer); err != nil {
-			logger.Warnf("Failed to add peer %s: %v", peer.PublicKey, err)
-			continue
-		}
-	}
-
-	// Save file
-	if err := iniFile.SaveTo(path); err != nil {
-		return fmt.Errorf("save config: %w", err)
-	}
-
-	// Remove backticks if PostUp/PostDown were added
-	if includePostUpDown {
-		return cleanBackticks(path)
-	}
-
-	return nil
-}
-
-func addPeerSection(iniFile *ini.File, peer ipc.PeerConfig) error {
-	psec, err := iniFile.NewSection("Peer")
-	if err != nil {
-		return err
-	}
-
-	setKey(psec, "PublicKey", peer.PublicKey)
-	setKeyIfNotEmpty(psec, "PresharedKey", peer.PresharedKey)
-	setKeyIfNotEmpty(psec, "AllowedIPs", strings.Join(peer.AllowedIPs, ","))
-	setKeyIfNotEmpty(psec, "Endpoint", peer.Endpoint)
-	setKeyIfPositive(psec, "PersistentKeepalive", peer.PersistentKeepalive)
-
-	return nil
-}
-
-// --- Peer Management ---
-func ExportPeerConfig(interfaceName string, peer ipc.PeerConfig, ifaceCfg InterfaceConfig, publicIP string, peerNumber int, dnsOverride string) (string, error) {
-
-	// Ensure peer directory exists
-	peerDir := peerDirPath(interfaceName)
-	if err := os.MkdirAll(peerDir, 0700); err != nil {
-		return "", fmt.Errorf("create peer dir: %w", err)
-	}
-
-	peerPath := filepath.Join(peerDir, fmt.Sprintf("Peer%d.conf", peerNumber))
-	iniFile := ini.Empty()
-
-	// Create Interface section for peer
-	ifSec, err := iniFile.NewSection("Interface")
-	if err != nil {
-		return "", fmt.Errorf("create interface section: %w", err)
-	}
-
-	// Set peer interface configuration
-	if len(peer.AllowedIPs) > 0 {
-		setKey(ifSec, "Address", peer.AllowedIPs[0])
-	}
-	setKeyIfPositive(ifSec, "ListenPort", ifaceCfg.ListenPort)
-
-	if peer.PrivateKey == "" {
-		return "", fmt.Errorf("peer private key is empty")
-	}
-	setKey(ifSec, "PrivateKey", peer.PrivateKey)
-
-	// DNS precedence: interface DNS > override > none
-	dnsVal := ""
-	if len(ifaceCfg.DNS) > 0 {
-		dnsVal = strings.Join(ifaceCfg.DNS, ",")
-	} else if dnsOverride != "" {
-		dnsVal = dnsOverride
-	}
-	setKeyIfNotEmpty(ifSec, "DNS", dnsVal)
-
-	// Create Peer section (connecting to server)
-	peerSec, err := iniFile.NewSection("Peer")
-	if err != nil {
-		return "", fmt.Errorf("create peer section: %w", err)
-	}
-
-	// Get server public key
-	serverKey, err := wgtypes.ParseKey(ifaceCfg.PrivateKey)
-	if err != nil {
-		return "", fmt.Errorf("parse server key: %w", err)
-	}
-
-	setKey(peerSec, "PublicKey", serverKey.PublicKey().String())
-	setKey(peerSec, "AllowedIPs", "0.0.0.0/1, 128.0.0.0/1, ::/0")
-
-	// Set endpoint if we have public IP - using simple format
-	if publicIP != "" && ifaceCfg.ListenPort > 0 {
-		endpoint := fmt.Sprintf("%s:%d", publicIP, ifaceCfg.ListenPort)
-		setKey(peerSec, "Endpoint", endpoint)
-	}
-
-	setKeyIfNotEmpty(peerSec, "PresharedKey", peer.PresharedKey)
-	setKeyIfPositive(peerSec, "PersistentKeepalive", peer.PersistentKeepalive)
-
-	// Save peer config
-	if err := iniFile.SaveTo(peerPath); err != nil {
-		return "", fmt.Errorf("save peer config: %w", err)
-	}
-
-	return peerPath, nil
-}
-
-// --- Network Helpers ---
-func isInterfaceUp(name string) bool {
-	return exec.Command("wg", "show", name).Run() == nil
-}
-
-func parseCSV(s string) []string {
-	if s == "" {
-		return nil
-	}
-
-	parts := strings.Split(s, ",")
-	result := make([]string, 0, len(parts))
-	for _, p := range parts {
-		if p = strings.TrimSpace(p); p != "" {
-			result = append(result, p)
-		}
-	}
-	return result
-}
-
-func isPeerSection(name string) bool {
-	return name == "Peer" || strings.HasPrefix(name, "Peer ")
-}
-
-// --- IP Address Management ---
-type ipManager struct {
-	netBase    net.IP
-	serverHost int
-}
-
-func newIPManager(serverCIDR string) (*ipManager, error) {
-	normalized, serverHost, err := normalizeCIDRv4Host(serverCIDR)
-	if err != nil {
-		return nil, err
-	}
-
-	netBase, err := ipv4NetBase(normalized)
-	if err != nil {
-		return nil, err
-	}
-
-	return &ipManager{
-		netBase:    netBase,
-		serverHost: serverHost,
-	}, nil
-}
-
-func (m *ipManager) findNextAvailable(peers []ipc.PeerConfig) (string, int, error) {
-	used := m.buildUsedIPMap(peers)
-
-	for i := minHostIP; i <= maxHostIP; i++ {
-		if !used[i] {
-			return m.makeIP(i), i, nil
-		}
-	}
-
-	return "", 0, fmt.Errorf("no available IPs in subnet")
-}
-
-func (m *ipManager) buildUsedIPMap(peers []ipc.PeerConfig) map[int]bool {
-	used := map[int]bool{
-		0:   true, // network
-		255: true, // broadcast
-	}
-
-	if m.serverHost > 0 {
-		used[m.serverHost] = true
-	}
-
-	for _, p := range peers {
-		for _, ip := range p.AllowedIPs {
-			if host := extractHostOctet(ip); host > 0 {
-				used[host] = true
-			}
-		}
-	}
-
-	return used
-}
-
-func (m *ipManager) makeIP(host int) string {
-	return fmt.Sprintf("%d.%d.%d.%d/32", m.netBase[0], m.netBase[1], m.netBase[2], host)
-}
-
-func extractHostOctet(ip string) int {
-	parts := strings.Split(ip, ".")
-	if len(parts) != 4 {
-		return -1
-	}
-
-	last := parts[3]
-	if idx := strings.Index(last, "/"); idx != -1 {
-		last = last[:idx]
-	}
-
-	host, err := strconv.Atoi(last)
-	if err != nil {
-		return -1
-	}
-	return host
-}
-
 // --- Handler Implementations ---
 func ListInterfaces([]string) (any, error) {
-	logger.Debugf("Listing interfaces")
+	logger.Debugf("ListInterfaces: listing interfaces")
 
 	pattern := filepath.Join(wgConfigDir, "*"+configExt)
 	files, err := filepath.Glob(pattern)
 	if err != nil {
+		logger.Errorf("ListInterfaces: glob failed for %s: %v", pattern, err)
 		return nil, fmt.Errorf("list interfaces: %w", err)
 	}
 
@@ -374,7 +35,7 @@ func ListInterfaces([]string) (any, error) {
 		name := strings.TrimSuffix(filepath.Base(f), configExt)
 		cfg, err := ParseWireGuardConfig(f)
 		if err != nil {
-			logger.Warnf("Failed to parse %s: %v", name, err)
+			logger.Warnf("ListInterfaces: failed to parse %s: %v", name, err)
 			continue
 		}
 
@@ -392,12 +53,13 @@ func ListInterfaces([]string) (any, error) {
 		})
 	}
 
-	logger.Infof("Found %d interfaces", len(interfaces))
+	logger.Infof("ListInterfaces: found %d interfaces", len(interfaces))
 	return interfaces, nil
 }
 
 func AddInterface(args []string) (any, error) {
 	if len(args) < 4 {
+		logger.Errorf("AddInterface: invalid arguments: %v", args)
 		return nil, fmt.Errorf("usage: add_interface <name> <addresses> <listenPort> <egressNic> [dns] [mtu] [peers_json] [numPeers]")
 	}
 
@@ -417,6 +79,7 @@ func AddInterface(args []string) (any, error) {
 		var err error
 		peers, err = generatePeers(addresses[0], numPeers)
 		if err != nil {
+			logger.Errorf("AddInterface: generate peers failed: %v", err)
 			return nil, fmt.Errorf("generate peers: %w", err)
 		}
 	}
@@ -424,6 +87,7 @@ func AddInterface(args []string) (any, error) {
 	// Generate server keys
 	privKey, err := wgtypes.GeneratePrivateKey()
 	if err != nil {
+		logger.Errorf("AddInterface: generate private key failed: %v", err)
 		return nil, fmt.Errorf("generate private key: %w", err)
 	}
 
@@ -438,29 +102,37 @@ func AddInterface(args []string) (any, error) {
 
 	// Get public IP once
 	publicIP, _ := utils.GetPublicIP()
+	if publicIP == "" {
+		logger.Warnf("AddInterface: GetPublicIP returned empty string")
+	}
 
 	// Prefer the gateway of the selected egress nic
 	gatewayDNS, _ := getGatewayForInterfaceIPv4(egressNic)
+	if gatewayDNS == "" {
+		logger.Warnf("AddInterface: could not determine gateway DNS for %s", egressNic)
+	}
 
 	// Export peer configs
 	for _, peer := range peers {
 		peerNumber := extractHostOctet(peer.AllowedIPs[0])
 		if _, err := ExportPeerConfig(name, peer, cfg, publicIP, peerNumber, gatewayDNS); err != nil {
-			logger.Warnf("Failed to export config for peer %s: %v", peer.PublicKey, err)
+			logger.Warnf("AddInterface: failed to export config for peer %s: %v", peer.PublicKey, err)
 		}
 	}
 
 	// Write main config
 	if err := WriteWireGuardConfigWithPostUpDown(configPath(name), cfg, egressNic); err != nil {
+		logger.Errorf("AddInterface: write config failed for %s: %v", name, err)
 		return nil, fmt.Errorf("write config: %w", err)
 	}
 
 	// Bring up interface
 	if _, err := UpInterface([]string{name}); err != nil {
+		logger.Errorf("AddInterface: failed to bring up %s: %v", name, err)
 		return nil, err
 	}
 
-	logger.Infof("Interface %s created and brought up", name)
+	logger.Infof("AddInterface: interface %s created and brought up", name)
 
 	return map[string]any{
 		"status":      "created",
@@ -472,56 +144,77 @@ func AddInterface(args []string) (any, error) {
 
 func RemoveInterface(args []string) (any, error) {
 	if len(args) < 1 {
+		logger.Errorf("RemoveInterface: invalid arguments: %v", args)
 		return nil, fmt.Errorf("usage: remove_interface <name>")
 	}
 
 	name := args[0]
-	logger.Infof("Removing interface: %s", name)
+	logger.Infof("RemoveInterface: removing interface %s", name)
 
-	// Try to bring down interface (ignore errors)
-	DownInterface([]string{name})
+	// Best-effort: bring it down, but don't abort on failure.
+	if _, err := DownInterface([]string{name}); err != nil {
+		logger.Warnf("RemoveInterface: failed to bring down %s; continuing with removal: %v", name, err)
+	} else {
+		logger.Infof("RemoveInterface: interface %s brought down before removal", name)
+	}
 
 	// Remove config file
-	if err := os.Remove(configPath(name)); err != nil {
-		return nil, fmt.Errorf("remove config: %w", err)
+	cfgPath := configPath(name)
+	if err := os.Remove(cfgPath); err != nil {
+		if os.IsNotExist(err) {
+			logger.Warnf("RemoveInterface: config file already missing for %s (%s)", name, cfgPath)
+		} else {
+			logger.Errorf("RemoveInterface: failed to remove config for %s (%s): %v", name, cfgPath, err)
+			return nil, fmt.Errorf("remove config: %w", err)
+		}
+	} else {
+		logger.Infof("RemoveInterface: removed config file for %s (%s)", name, cfgPath)
 	}
 
-	// Remove peer directory
+	// Remove peer directory (best-effort)
 	peerDir := peerDirPath(name)
 	if err := os.RemoveAll(peerDir); err != nil && !os.IsNotExist(err) {
-		logger.Warnf("Could not remove peer dir %s: %v", peerDir, err)
+		logger.Warnf("RemoveInterface: could not remove peer dir %s: %v", peerDir, err)
+	} else {
+		logger.Infof("RemoveInterface: removed peer directory for %s (%s)", name, peerDir)
 	}
 
-	logger.Infof("Interface %s removed", name)
+	logger.Infof("RemoveInterface: interface %s removed", name)
 	return "removed", nil
 }
 
 func AddPeer(args []string) (any, error) {
 	if len(args) < 1 {
+		logger.Errorf("AddPeer: invalid arguments: %v", args)
 		return nil, fmt.Errorf("usage: add_peer <interface>")
 	}
 
 	interfaceName := args[0]
+	logger.Infof("AddPeer: adding peer to %s", interfaceName)
 
 	// Read current config
 	cfg, err := ParseWireGuardConfig(configPath(interfaceName))
 	if err != nil {
+		logger.Errorf("AddPeer: read config failed for %s: %v", interfaceName, err)
 		return nil, fmt.Errorf("read config: %w", err)
 	}
 
 	if len(cfg.Address) == 0 {
+		logger.Errorf("AddPeer: interface %s has no address", interfaceName)
 		return nil, fmt.Errorf("interface has no address")
 	}
 
 	// Initialize IP manager
 	ipMgr, err := newIPManager(cfg.Address[0])
 	if err != nil {
+		logger.Errorf("AddPeer: init IP manager failed: %v", err)
 		return nil, fmt.Errorf("init IP manager: %w", err)
 	}
 
 	// Find next available IP
 	nextIP, peerNumber, err := ipMgr.findNextAvailable(cfg.Peers)
 	if err != nil {
+		logger.Errorf("AddPeer: find next available IP failed: %v", err)
 		return nil, err
 	}
 
@@ -536,19 +229,37 @@ func AddPeer(args []string) (any, error) {
 		Name:                fmt.Sprintf("Peer%d", peerNumber),
 	}
 
-	// Update config
+	// Update config (write to disk)
 	cfg.Peers = append(cfg.Peers, peer)
 	if err := WriteWireGuardConfig(configPath(interfaceName), cfg); err != nil {
+		logger.Errorf("AddPeer: write config failed for %s: %v", interfaceName, err)
 		return nil, fmt.Errorf("write config: %w", err)
 	}
 
 	// Export peer config
 	publicIP, _ := utils.GetPublicIP()
+	if publicIP == "" {
+		logger.Warnf("AddPeer: GetPublicIP returned empty string")
+	}
 	gatewayDNS, _ := getDefaultGatewayIPv4()
+	if gatewayDNS == "" {
+		logger.Warnf("AddPeer: could not determine default gateway DNS")
+	}
 	if _, err := ExportPeerConfig(interfaceName, peer, cfg, publicIP, peerNumber, gatewayDNS); err != nil {
+		logger.Errorf("AddPeer: export peer config failed: %v", err)
 		return nil, fmt.Errorf("export peer config: %w", err)
 	}
 
+	// Bounce interface to apply changes
+	if _, err := DownInterface([]string{interfaceName}); err != nil {
+		logger.Warnf("AddPeer: wg-quick down %s failed (continuing): %v", interfaceName, err)
+	}
+	if _, err := UpInterface([]string{interfaceName}); err != nil {
+		logger.Errorf("AddPeer: wg-quick up %s failed after adding peer: %v", interfaceName, err)
+		return nil, fmt.Errorf("restart interface: %w", err)
+	}
+
+	logger.Infof("AddPeer: added %s with IP %s to %s and restarted interface", peer.Name, nextIP, interfaceName)
 	return map[string]any{
 		"peer_name":  peer.Name,
 		"public_key": peer.PublicKey,
@@ -558,11 +269,13 @@ func AddPeer(args []string) (any, error) {
 
 func RemovePeerByName(args []string) (any, error) {
 	if len(args) < 2 {
+		logger.Errorf("RemovePeerByName: invalid arguments: %v", args)
 		return nil, fmt.Errorf("usage: remove_peer <interface> <peer>")
 	}
 
 	interfaceName := args[0]
 	peerName := args[1]
+	logger.Infof("RemovePeerByName: removing peer %s from %s", peerName, interfaceName)
 
 	// Read peer config to get AllowedIP
 	peerPath := peerConfigPath(interfaceName, peerName)
@@ -570,17 +283,20 @@ func RemovePeerByName(args []string) (any, error) {
 	// Extract Address from peer config
 	iniFile, err := ini.Load(peerPath)
 	if err != nil {
+		logger.Errorf("RemovePeerByName: parse peer config %s failed: %v", peerPath, err)
 		return nil, fmt.Errorf("parse peer config: %w", err)
 	}
 
 	allowedIP := iniFile.Section("Interface").Key("Address").String()
 	if allowedIP == "" {
+		logger.Errorf("RemovePeerByName: peer config %s missing Address", peerPath)
 		return nil, fmt.Errorf("peer config missing Address")
 	}
 
 	// Read main config
 	cfg, err := ParseWireGuardConfig(configPath(interfaceName))
 	if err != nil {
+		logger.Errorf("RemovePeerByName: read main config for %s failed: %v", interfaceName, err)
 		return nil, fmt.Errorf("read main config: %w", err)
 	}
 
@@ -603,6 +319,7 @@ func RemovePeerByName(args []string) (any, error) {
 	}
 
 	if !found {
+		logger.Errorf("RemovePeerByName: peer with IP %s not found in %s", allowedIP, interfaceName)
 		return nil, fmt.Errorf("peer not found with IP %s", allowedIP)
 	}
 
@@ -610,26 +327,34 @@ func RemovePeerByName(args []string) (any, error) {
 
 	// Write updated config
 	if err := WriteWireGuardConfig(configPath(interfaceName), cfg); err != nil {
+		logger.Errorf("RemovePeerByName: write updated config failed: %v", err)
 		return nil, fmt.Errorf("write config: %w", err)
 	}
 
 	// Remove peer config file
 	if err := os.Remove(peerPath); err != nil && !os.IsNotExist(err) {
-		logger.Warnf("Could not remove peer config: %v", err)
+		logger.Warnf("RemovePeerByName: could not remove peer config %s: %v", peerPath, err)
+	} else {
+		logger.Infof("RemovePeerByName: removed peer config file %s", peerPath)
 	}
 
+	logger.Infof("RemovePeerByName: peer %s removed from %s", peerName, interfaceName)
 	return "removed", nil
 }
 
 func ListPeers(args []string) (any, error) {
 	if len(args) < 1 {
+		logger.Errorf("ListPeers: invalid arguments: %v", args)
 		return nil, fmt.Errorf("usage: list_exported_peers <interface>")
 	}
 
 	interfaceName := args[0]
+	logger.Debugf("ListPeers: listing exported peers for %s", interfaceName)
+
 	pattern := filepath.Join(peerDirPath(interfaceName), "*"+configExt)
 	files, err := filepath.Glob(pattern)
 	if err != nil {
+		logger.Errorf("ListPeers: glob failed for %s: %v", pattern, err)
 		return nil, fmt.Errorf("list peer configs: %w", err)
 	}
 
@@ -638,7 +363,7 @@ func ListPeers(args []string) (any, error) {
 	for _, file := range files {
 		iniFile, err := ini.Load(file)
 		if err != nil {
-			logger.Warnf("Failed to parse %s: %v", file, err)
+			logger.Warnf("ListPeers: failed to parse %s: %v", file, err)
 			continue
 		}
 
@@ -662,11 +387,13 @@ func ListPeers(args []string) (any, error) {
 		peers = append(peers, pc)
 	}
 
+	logger.Infof("ListPeers: found %d exported peers for %s", len(peers), interfaceName)
 	return peers, nil
 }
 
 func UpInterface(args []string) (any, error) {
 	if len(args) < 1 {
+		logger.Errorf("UpInterface: invalid arguments: %v", args)
 		return nil, fmt.Errorf("usage: up_interface <name>")
 	}
 
@@ -675,11 +402,11 @@ func UpInterface(args []string) (any, error) {
 	out, err := cmd.CombinedOutput()
 
 	if err != nil {
-		logger.Errorf("Failed to bring up %s: %v (%s)", name, err, string(out))
+		logger.Errorf("UpInterface: failed to bring up %s: %v (%s)", name, err, string(out))
 		return nil, fmt.Errorf("bring up interface: %w", err)
 	}
 
-	logger.Infof("Interface %s brought up", name)
+	logger.Infof("UpInterface: interface %s brought up", name)
 	return map[string]any{
 		"status": "on",
 		"output": string(out),
@@ -688,6 +415,7 @@ func UpInterface(args []string) (any, error) {
 
 func DownInterface(args []string) (any, error) {
 	if len(args) < 1 {
+		logger.Errorf("DownInterface: invalid arguments: %v", args)
 		return nil, fmt.Errorf("usage: down_interface <name>")
 	}
 
@@ -696,11 +424,11 @@ func DownInterface(args []string) (any, error) {
 	out, err := cmd.CombinedOutput()
 
 	if err != nil {
-		logger.Errorf("Failed to bring down %s: %v (%s)", name, err, string(out))
+		logger.Errorf("DownInterface: failed to bring down %s: %v (%s)", name, err, string(out))
 		return nil, fmt.Errorf("bring down interface: %w", err)
 	}
 
-	logger.Infof("Interface %s brought down", name)
+	logger.Infof("DownInterface: interface %s brought down", name)
 	return map[string]any{
 		"status": "off",
 		"output": string(out),
@@ -709,254 +437,67 @@ func DownInterface(args []string) (any, error) {
 
 func PeerQRCode(args []string) (any, error) {
 	if len(args) < 2 {
+		logger.Errorf("PeerQRCode: invalid arguments: %v", args)
 		return nil, fmt.Errorf("usage: peer_qrcode <interface> <peername>")
 	}
 
 	peerPath := peerConfigPath(args[0], args[1])
 	rawConfig, err := os.ReadFile(peerPath)
 	if err != nil {
+		logger.Errorf("PeerQRCode: read peer config %s failed: %v", peerPath, err)
 		return nil, fmt.Errorf("read peer config: %w", err)
 	}
 
 	// Generate QR code
 	png, err := qrcode.Encode(string(rawConfig), qrcode.Medium, 256)
 	if err != nil {
+		logger.Errorf("PeerQRCode: QR encode failed for %s: %v", peerPath, err)
 		return nil, fmt.Errorf("generate QR code: %w", err)
 	}
 
 	dataURI := "data:image/png;base64," + base64.StdEncoding.EncodeToString(png)
+	logger.Infof("PeerQRCode: generated QR for %s", peerPath)
 	return map[string]string{"qrcode": dataURI}, nil
 }
 
 func PeerConfigDownload(args []string) (any, error) {
 	if len(args) < 2 {
+		logger.Errorf("PeerConfigDownload: invalid arguments: %v", args)
 		return nil, fmt.Errorf("usage: peer_config_download <interface> <peername>")
 	}
 
 	peerPath := peerConfigPath(args[0], args[1])
 	data, err := os.ReadFile(peerPath)
 	if err != nil {
+		logger.Errorf("PeerConfigDownload: read peer config %s failed: %v", peerPath, err)
 		return nil, fmt.Errorf("read peer config: %w", err)
 	}
 
+	logger.Infof("PeerConfigDownload: served %d bytes from %s", len(data), peerPath)
 	return map[string]string{"config": string(data)}, nil
 }
 
 func GetKeys(args []string) (any, error) {
 	if len(args) < 1 {
+		logger.Errorf("GetKeys: invalid arguments: %v", args)
 		return nil, fmt.Errorf("usage: get_keys <interface>")
 	}
 
 	cfg, err := ParseWireGuardConfig(configPath(args[0]))
 	if err != nil {
+		logger.Errorf("GetKeys: read config failed for %s: %v", args[0], err)
 		return nil, fmt.Errorf("read config: %w", err)
 	}
 
 	key, err := wgtypes.ParseKey(cfg.PrivateKey)
 	if err != nil {
+		logger.Errorf("GetKeys: parse private key failed: %v", err)
 		return nil, fmt.Errorf("parse key: %w", err)
 	}
 
+	logger.Infof("GetKeys: provided keys for %s", args[0])
 	return map[string]string{
 		"private_key": cfg.PrivateKey,
 		"public_key":  key.PublicKey().String(),
 	}, nil
-}
-
-// --- Helper functions ---
-func normalizeAddresses(addrs []string) []string {
-	result := make([]string, 0, len(addrs))
-	for _, addr := range addrs {
-		normalized, _, err := normalizeCIDRv4Host(addr)
-		if err != nil {
-			result = append(result, addr) // Keep original if not IPv4 CIDR
-		} else {
-			result = append(result, normalized)
-		}
-	}
-	return result
-}
-
-func parseOptionalCSV(args []string, index int) []string {
-	if index < len(args) && args[index] != "" && args[index] != "null" {
-		return parseCSV(args[index])
-	}
-	return nil
-}
-
-func parseOptionalInt(args []string, index int, defaultVal int) int {
-	if index < len(args) && args[index] != "" && args[index] != "null" {
-		val, _ := strconv.Atoi(args[index])
-		return val
-	}
-	return defaultVal
-}
-
-func parseOptionalPeers(args []string, index int) []ipc.PeerConfig {
-	if index < len(args) && args[index] != "" && args[index] != "null" && args[index] != "[]" {
-		var peers []ipc.PeerConfig
-		if err := json.Unmarshal([]byte(args[index]), &peers); err == nil {
-			return peers
-		}
-	}
-	return nil
-}
-
-func generatePeers(serverAddr string, count int) ([]ipc.PeerConfig, error) {
-	ipMgr, err := newIPManager(serverAddr)
-	if err != nil {
-		return nil, err
-	}
-
-	peers := make([]ipc.PeerConfig, 0, count)
-	used := ipMgr.buildUsedIPMap(nil)
-
-	for i := 0; i < count; i++ {
-		// Find next available IP
-		var peerIP string
-		var host int
-		for h := minHostIP; h <= maxHostIP; h++ {
-			if !used[h] {
-				peerIP = ipMgr.makeIP(h)
-				host = h
-				used[h] = true
-				break
-			}
-		}
-
-		if peerIP == "" {
-			return nil, fmt.Errorf("insufficient IPs for %d peers", count)
-		}
-
-		// Generate keys
-		privKey, err := wgtypes.GeneratePrivateKey()
-		if err != nil {
-			return nil, fmt.Errorf("generate key for peer %d: %w", i+1, err)
-		}
-
-		peers = append(peers, ipc.PeerConfig{
-			PublicKey:           privKey.PublicKey().String(),
-			PrivateKey:          privKey.String(),
-			AllowedIPs:          []string{peerIP},
-			PersistentKeepalive: defaultKeepalive,
-			Name:                fmt.Sprintf("Peer%d", host),
-		})
-	}
-
-	return peers, nil
-}
-
-// --- INI Helper Functions ---
-func setKey(section *ini.Section, key, value string) {
-	if _, err := section.NewKey(key, value); err != nil {
-		logger.Warnf("Failed to set %s: %v", key, err)
-	}
-}
-
-func setKeyIfNotEmpty(section *ini.Section, key, value string) {
-	if value != "" {
-		setKey(section, key, value)
-	}
-}
-
-func setKeyIfPositive(section *ini.Section, key string, value int) {
-	if value > 0 {
-		setKey(section, key, strconv.Itoa(value))
-	}
-}
-
-func cleanBackticks(path string) error {
-	data, err := os.ReadFile(path)
-	if err != nil {
-		return err
-	}
-
-	cleaned := strings.ReplaceAll(string(data), "`", "")
-	return os.WriteFile(path, []byte(cleaned), 0600)
-}
-
-// --- IPv4 Helper Functions ---
-func normalizeCIDRv4Host(cidr string) (string, int, error) {
-	ip, ipNet, err := net.ParseCIDR(strings.TrimSpace(cidr))
-	if err != nil {
-		return cidr, 0, err
-	}
-
-	v4 := ip.To4()
-	if v4 == nil {
-		return cidr, 0, nil // Not IPv4
-	}
-
-	// If IP is the network address, increment to first usable host
-	if v4.Equal(ipNet.IP.To4()) {
-		host := make(net.IP, len(v4))
-		copy(host, v4)
-		host[3]++
-		ones, _ := ipNet.Mask.Size()
-		return fmt.Sprintf("%s/%d", host.String(), ones), int(host[3]), nil
-	}
-
-	// Otherwise keep as-is
-	ones, _ := ipNet.Mask.Size()
-	return fmt.Sprintf("%s/%d", v4.String(), ones), int(v4[3]), nil
-}
-
-func ipv4NetBase(cidr string) (net.IP, error) {
-	_, ipNet, err := net.ParseCIDR(strings.TrimSpace(cidr))
-	if err != nil {
-		return nil, err
-	}
-	return ipNet.IP.To4(), nil
-}
-
-// getDefaultGatewayIPv4 reads /proc/net/route and returns the default gateway.
-func getDefaultGatewayIPv4() (string, error) {
-	return getGatewayFromRouteFile("")
-}
-
-// getGatewayForInterfaceIPv4 returns the default gateway for a specific interface.
-func getGatewayForInterfaceIPv4(iface string) (string, error) {
-	return getGatewayFromRouteFile(iface)
-}
-
-func getGatewayFromRouteFile(matchIface string) (string, error) {
-	data, err := os.ReadFile("/proc/net/route")
-	if err != nil {
-		return "", fmt.Errorf("read /proc/net/route: %w", err)
-	}
-	lines := strings.Split(strings.TrimSpace(string(data)), "\n")
-	// fields: Iface Destination Gateway Flags RefCnt Use Metric Mask MTU Window IRTT
-	for i := 1; i < len(lines); i++ {
-		fields := strings.Fields(lines[i])
-		if len(fields) < 3 {
-			continue
-		}
-		iface := fields[0]
-		dest := fields[1]
-		gwHex := fields[2]
-
-		// Only default route (Destination == 00000000)
-		if dest != "00000000" {
-			continue
-		}
-		if matchIface != "" && iface != matchIface {
-			continue
-		}
-		gwIP, err := hexLEToIPv4(gwHex)
-		if err != nil || gwIP == "0.0.0.0" {
-			continue
-		}
-		return gwIP, nil
-	}
-	return "", fmt.Errorf("default gateway not found")
-}
-
-func hexLEToIPv4(hexStr string) (string, error) {
-	u, err := strconv.ParseUint(hexStr, 16, 32)
-	if err != nil {
-		return "", err
-	}
-	var b [4]byte
-	binary.LittleEndian.PutUint32(b[:], uint32(u))
-	return net.IPv4(b[0], b[1], b[2], b[3]).String(), nil
 }
